rename data mover files
