+++ src/gen_ai/genai_explore.py
@@ -53,7 +53,7 @@ conversation = ConversationChain(
 )
 
 
-from src.data.csv_mover import CSVMover
+from src.data.data_mover import CSVMover
 from src.data.recipient_mapper import RecipientMapper
 from src.data.data_wrangling import DateTimeConverter
 from src.data.recipient_mapper import RecipientMapper
